[feature/slices] Players slice
