+++ src/store/game/players/tests/set-next-player.test.ts
@@ -1,0 +1,29 @@
+import { describe, expect, it } from "vitest";
+import { DPlayers } from "@/data/players.data";
+import { createTestStore } from "./setup/create-test-store";
+
+describe("setNextPlayer()", () => {
+  it("should set next player", () => {
+    const store = createTestStore();
+    store.setState({
+      players: DPlayers,
+      currentPlayerIndex: 0,
+    });
+
+    store.getState().setNextPlayer();
+
+    expect(store.getState().currentPlayerIndex).toBe(1);
+  });
+
+  it("should loop back to first player", () => {
+    const store = createTestStore();
+    store.setState({
+      players: DPlayers,
+      currentPlayerIndex: DPlayers.length - 1,
+    });
+
+    store.getState().setNextPlayer();
+
+    expect(store.getState().currentPlayerIndex).toBe(0);
+  });
+});
