+++ src/store/game/players/tests/set-player-resources.test.ts
@@ -1,0 +1,20 @@
+import { expect, it } from "vitest";
+import { DPlayers } from "@/data/players.data";
+import { createTestStore } from "./setup/create-test-store";
+
+it("setPlayerResources()", () => {
+  const store = createTestStore();
+  store.setState({
+    players: DPlayers,
+    currentPlayerIndex: 0,
+  });
+
+  store.getState().setPlayerResources(1, { stone: 5, wood: 2 });
+  const updatedPlayer = store
+    .getState()
+    .players.find((player) => player.id === 1);
+
+  expect(updatedPlayer).toBeDefined();
+  expect(updatedPlayer?.resources.stone).toBe(5);
+  expect(updatedPlayer?.resources.wood).toBe(2);
+});
