+++ src/store/game/players/tests/setup/create-test-store.ts
@@ -1,0 +1,8 @@
+import { create } from "zustand";
+import { IPlayersSlice } from "../../players.types";
+import { createPlayersSlice } from "../..";
+
+export const createTestStore = () =>
+  create<IPlayersSlice>((...args) => ({
+    ...createPlayersSlice(...args),
+  }));
